Improve data types and consistency; behave sanely for static credentials
The bulk of AbstractProvider is written with refreshing in mind. However, when
moving towards subclassing StaticProvider it was clear that certain behavior
needed to be shortcutted with certain fields missing. This commit makes those
fields optional and rearranges the method types to allow the user to supply
more information than necessary if desired, but require the minimum information
necessary for each particular method. It also adds some more consistency and
data-checking, to ensure that the types we've defined are always true.

Significantly, this commit introduces a hydration step which, currently, will
always run since one of the pieces of information (expires_in) is never truly
available without doing some shenanigans. This required additional mocks in
the tests and some shuffling-around, but the outcome is that consistent data
is now always available.

No matter what the user inputs for scopes, the scopes the class actually cares
about will always be what Twitch says they are. No matter what the user adds
for the "initial expiry" time, it will always be as correct as possible too.